CHANGE  - moved `CryptoStrategy` to separate file  - moved SimpleEncryptor into folder
